char is always bigger than 0.

+++ org.eclipse.titan.designer/src/org/eclipse/titan/designer/AST/TTCN3/types/CharString_Type.java
@@ -339,7 +339,7 @@ public final class CharString_Type extends Type {
 		int i = 0;
 		while (i < length) {
 			final char value = defaultValue.charAt(i);
-			if (value < 0) {
+			if ((byte)value < 0) {
 				getLocation().reportSemanticError(MessageFormat.format("Invalid {0} JSON default value", getTypename()));
 				return;
 			}
